+++ .travis.yml
@@ -2,6 +2,6 @@ language: python
 install:
   - pip install coveralls
 script:
-  nosetests --with-coverage tex2Wiki DLMF_preprocessing maple2latex
+  nosetests --with-coverage tex2Wiki DLMF_preprocessing maple2latex main_page
 after_success:
   coveralls
