+++ main_page/test/update_headers_tests/res.1.txt
@@ -1,0 +1,45 @@
+drmf_bof
+'''Main Page'''
+drmf_eof
+drmf_bof
+'''Definition:Definition1'''
+<div id="drmf_head">
+<div id="alignleft"> << [[Definition:Definition3|Definition:Definition3]] </div>
+<div id="aligncenter"> [[Main_Page|Main Page]] </div>
+<div id="alignright"> [[Definition:Definition2|Definition:Definition2]] >> </div>
+</div>
+Definition1 Text.
+<br /><div id="drmf_foot">
+<div id="alignleft"> << [[Definition:Definition3|Definition:Definition3]] </div>
+<div id="aligncenter"> [[Main_Page|Main Page]] </div>
+<div id="alignright"> [[Definition:Definition2|Definition:Definition2]] >> </div>
+</div>
+drmf_eof
+drmf_bof
+'''Definition:Definition2'''
+<div id="drmf_head">
+<div id="alignleft"> << [[Definition:Definition1|Definition:Definition1]] </div>
+<div id="aligncenter"> [[Main_Page|Main Page]] </div>
+<div id="alignright"> [[Definition:Definition3|Definition:Definition3]] >> </div>
+</div>
+Definition2 Text.
+<br /><div id="drmf_foot">
+<div id="alignleft"> << [[Definition:Definition1|Definition:Definition1]] </div>
+<div id="aligncenter"> [[Main_Page|Main Page]] </div>
+<div id="alignright"> [[Definition:Definition3|Definition:Definition3]] >> </div>
+</div>
+drmf_eof
+drmf_bof
+'''Definition:Definition3'''
+<div id="drmf_head">
+<div id="alignleft"> << [[Definition:Definition2|Definition:Definition2]] </div>
+<div id="aligncenter"> [[Main_Page|Main Page]] </div>
+<div id="alignright"> [[Definition:Definition1|Definition:Definition1]] >> </div>
+</div>
+Definition3 Text.
+<br /><div id="drmf_foot">
+<div id="alignleft"> << [[Definition:Definition2|Definition:Definition2]] </div>
+<div id="aligncenter"> [[Main_Page|Main Page]] </div>
+<div id="alignright"> [[Definition:Definition1|Definition:Definition1]] >> </div>
+</div>
+drmf_eof
